# use struct slice to improve performance