--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -30,6 +30,9 @@ jobs:
     - name: Test
       run: go test -v .
 
+    - name: Benchmark
+      run: go test -bench=.
+
     - name: Vet
       run: go vet -v .