--- a/int64_sequence_test.go
+++ b/int64_sequence_test.go
@@ -137,6 +137,9 @@ func TestInt64Sequence_Sort(t *testing.T) {
 		{
 			s: RandomInt64Sequence(10),
 		},
+		{
+			s: nil,
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {