```text
$ git checkout a83787d
$ go test -bench=. -count=10 | tee old.txt
goos: darwin
goarch: amd64
pkg: github.com/gochore/timeseq
BenchmarkFloat64Sequence_GC-12    	    5863	    190792 ns/op	   40201 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6499	    179948 ns/op	   40204 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6738	    178556 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6656	    178659 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6878	    177443 ns/op	   40204 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6490	    177874 ns/op	   40204 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6837	    177252 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6768	    177972 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6812	    178975 ns/op	   40204 B/op	    1001 allocs/op
BenchmarkFloat64Sequence_GC-12    	    6962	    177904 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6978	    172290 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6969	    174044 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6852	    172731 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6956	    172436 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6933	    173038 ns/op	   40205 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6933	    172670 ns/op	   40204 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6973	    172428 ns/op	   40204 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6886	    173817 ns/op	   40203 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6985	    173313 ns/op	   40204 B/op	    1001 allocs/op
BenchmarkInt64Sequence_GC-12      	    6842	    172281 ns/op	   40204 B/op	    1001 allocs/op
PASS
ok  	github.com/gochore/timeseq	27.573s
$ git checkout 8cce2c7
$ go test -bench=. -count=10 | tee new.txt
goos: darwin
goarch: amd64
pkg: github.com/gochore/timeseq
BenchmarkFloat64Sequence_GC-12    	    8206	    149801 ns/op	   32777 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    7933	    147796 ns/op	   32777 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8365	    148696 ns/op	   32776 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8155	    147133 ns/op	   32776 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8358	    144770 ns/op	   32777 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8427	    142049 ns/op	   32777 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8278	    141787 ns/op	   32778 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8637	    141518 ns/op	   32778 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8524	    141810 ns/op	   32777 B/op	       1 allocs/op
BenchmarkFloat64Sequence_GC-12    	    8424	    141487 ns/op	   32778 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8756	    137234 ns/op	   32777 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8854	    136242 ns/op	   32774 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8854	    136648 ns/op	   32774 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8692	    138555 ns/op	   32777 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8688	    140837 ns/op	   32780 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8464	    138793 ns/op	   32778 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8737	    138695 ns/op	   32778 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8702	    138487 ns/op	   32777 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8580	    138996 ns/op	   32776 B/op	       1 allocs/op
BenchmarkInt64Sequence_GC-12      	    8680	    142901 ns/op	   32776 B/op	       1 allocs/op
PASS
ok  	github.com/gochore/timeseq	24.871s
$ benchstat old.txt new.txt
name                   old time/op    new time/op    delta
Float64Sequence_GC-12     178µs ± 1%     145µs ± 4%  -18.85%  (p=0.000 n=9+10)
Int64Sequence_GC-12       173µs ± 1%     138µs ± 2%  -20.03%  (p=0.000 n=10+9)

name                   old alloc/op   new alloc/op   delta
Float64Sequence_GC-12    40.2kB ± 0%    32.8kB ± 0%  -18.47%  (p=0.000 n=9+10)
Int64Sequence_GC-12      40.2kB ± 0%    32.8kB ± 0%  -18.47%  (p=0.000 n=10+10)

name                   old allocs/op  new allocs/op  delta
Float64Sequence_GC-12     1.00k ± 0%     0.00k ± 0%  -99.90%  (p=0.000 n=10+10)
Int64Sequence_GC-12       1.00k ± 0%     0.00k ± 0%  -99.90%  (p=0.000 n=10+10)
```